+++ examples/platform/resources.h
@@ -25,10 +25,10 @@
 #include "defines.h"
 
 // VRAM offsets of background and sprite image data.
-extern uint16_t g_bg_offset;
-extern uint16_t g_moon_offset;
-extern uint16_t g_level_offset;
-extern uint16_t g_bat_offset;
+extern uint32_t g_bg_offset;
+extern uint32_t g_moon_offset;
+extern uint32_t g_level_offset;
+extern uint32_t g_bat_offset;
 extern uint16_t g_level_buffer;
 
 // Load image, palette, and tilemap data from file system.
